WIP - Add functionality to all the operations, but still untested.

+++ virtual-machine-v2/virtual-machine-v2/MemoryTableView.swift
@@ -10,7 +10,7 @@ import Foundation
 import UIKit
 
 final class MemoryTableView : UITableView {
-    var memory: [Int] = [] { didSet { handleMemoryChanged() } }
+    var memory: [Decimal] = [] { didSet { handleMemoryChanged() } }
 
     // MARK: Initialization
     override func awakeFromNib() {
